adding usage of limited swap
default of memory.swap.max is max. The OCI hook
will set 0 on every container. The wasp-agent main loop
will calculate the limit and will set it to the cgroup.

since same vars are used in few files, a common.sh created
for mutual import.

a cache cleanup function was introduced as part of trap callback.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>