functests: use same vesion of ginkgo for both CLI and e2e tests
We use vendored Ginkgo package to build the e2e tests. Ginkgo
CLI was also installed in the builder container from latest tag.
This introduced verion mismatch between the ginkgo CLI version
and the e2e code based on pinned vendored ginkgo version.

In order to reduce the boilerplate of maintaining the versions
in two different locations, just build the gikngo CLI from the
vendored package, and use this binary to build tests.test

In that case we don't need gikngo to be installed on the builder
container, and we also don't need to to rsync the vendored package
since after building ginkgo inside the build container there are
build leftovers in the vendor directory.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>