eviction-controller: decouple the taint logic
taint logic needs to be synced with the operator
that generates the wasp-agent ds manifest, because
wasp-agent needs to tolerate its own taint.

this is needed in case wasp-agent pod will be
restarted while the node is tainted. wasp-agent
needs to complete its eviction operation and remove
the taint when resources are back to normal.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>