manifests: fix machine configs for openshift deployment
there is a problem when user initiates systemd kubelet restart
the swap provision service also restarts, that causes failure
since the swap file already exists, and no need to provision it again.

In order to fix this added ConditionPathExists to swap provision unit
and the unit was divided into two units:
- swap provision unit runs conditionally based on existence of the swap file
- system slice configuration runs unconditionally in separate unit

both units need to be dependent on kubelet: swap provision requires
kubelet restart, and system slice configuration needs to be synced
with kubelet as well in order to prevent races.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>