render the OCI hook script from the CRIO default runtime
CRI-O package swithced to crun as its default runtime.
Nevertheless Openshift can use crio drop-in in order to preserve
runc as the default runtime, in order to preseve backward compatiblity.

wasp-agent needs to support both runc and crun, and the command to
update the swap max has different syntax.

this change proposes parsing the default runtime from crio config
and render the OCI hook script based on the runtime.

the OCI hook script is now generated in runtime depending on the crio
config. hook.sh turned into a Go template.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>